+++ Changelog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Added
 - Configuration.get_option() returns converted option value if an option was assigned a type (int, float, boolean). Available option types are also used for validating configuration. PluginConfiguration.init_option_types() is added to enable option typing in plugins. (#56)
+- The output of the 'list' command shows the difference of total earnings and expenses. (#29)
 ### Changed
 - The PluginConfiguration.validate() method is not required to be implemented in derived classes anymore.
 - Instead of having Configuration.get_option() return a section if no option was specified, a method get_section() is introduced.